duckdb-server/requirements.txt
blinker==1.9.0
click==8.3.1
duckdb==1.5.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.4
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
Werkzeug==3.1.7
